pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
    "sqlalchemy>=2.0.42",
    "werkzeug>=3.1.3",
    "paypalrestsdk>=1.13.3",
    "requests>=2.32.4",
    "trafilatura>=2.0.0",
    "selenium>=4.34.2",
    "python-magic>=0.4.27",
    "beautifulsoup4>=4.13.4",
    "schedule>=1.2.2",
    "stripe>=12.4.0",
    "flask-dance>=7.1.0",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
    "fake-useragent>=2.2.0",
    "psutil>=7.0.0",
    "cloudinary>=1.44.1",
    "stream-unzip>=0.0.99",
    "defusedxml>=0.7.1",
    "pymysql>=1.1.2",
    "psycopg2-binary>=2.9.10",
]
